server/requirements.txt
django
channels
djangorestframework
djangorestframework-simplejwt
channels-redis
psycopg2-binary
drf_yasg
drf-spectacular
locust # для нагрузочного тестирования
django-silk # профилировщик
Pillow